+++ .github/bootstrap-controller.sh
@@ -122,11 +122,29 @@ if [[ $? != 0 ]]; then echo ">>> ERROR: $scriptName:$CMD."; exit 1; fi
 #####################################################################################
 # Terraform
 
-# https://releases.hashicorp.com/terraform/0.13.4/terraform_0.13.4_linux_amd64.zip
-CMD="sudo wget -O ./terraform.zip https://releases.hashicorp.com/terraform/0.13.3/terraform_0.13.3_linux_amd64.zip";
+# # https://releases.hashicorp.com/terraform/0.13.4/terraform_0.13.4_linux_amd64.zip
+# CMD="sudo wget -O ./terraform.zip https://releases.hashicorp.com/terraform/0.13.3/terraform_0.13.3_linux_amd64.zip";
+# echo ">>> cmd: $CMD"; $CMD; if [[ $? != 0 ]]; then echo ">>> ERROR: $scriptName:$CMD."; exit 1; fi
+#
+# CMD="sudo unzip -o ./terraform.zip -d /usr/local/bin";
+# echo ">>> cmd: $CMD"; $CMD; if [[ $? != 0 ]]; then echo ">>> ERROR: $scriptName:$CMD."; exit 1; fi
+
+CMD="curl -fsSL https://apt.releases.hashicorp.com/gpg | sudo apt-key add -";
+echo ">>> cmd: $CMD";
+curl -fsSL https://apt.releases.hashicorp.com/gpg | sudo apt-key add -
+if [[ $? != 0 ]]; then echo ">>> ERROR: $scriptName:$CMD."; exit 1; fi
+
+lsbRelease=$(lsb_release -cs)
+repo="deb [arch=amd64] https://apt.releases.hashicorp.com $lsbRelease main"
+CMD="sudo apt-add-repository \"$repo\"";
+echo ">>> cmd: $CMD";
+sudo apt-add-repository "deb [arch=amd64] https://apt.releases.hashicorp.com $(lsb_release -cs) main"
+if [[ $? != 0 ]]; then echo ">>> ERROR: $scriptName:$CMD."; exit 1; fi
+
+CMD="sudo apt-get update";
 echo ">>> cmd: $CMD"; $CMD; if [[ $? != 0 ]]; then echo ">>> ERROR: $scriptName:$CMD."; exit 1; fi
 
-CMD="sudo unzip -o ./terraform.zip -d /usr/local/bin";
+CMD="sudo apt-get install terraform";
 echo ">>> cmd: $CMD"; $CMD; if [[ $? != 0 ]]; then echo ">>> ERROR: $scriptName:$CMD."; exit 1; fi
 
 #####################################################################################