+++ analytics/.devel/run.bg.sh
@@ -23,9 +23,10 @@ export TEST_RESULTS_DIR="$projectHome/uc-non-persistent/test-results/stats"
 export ANALYSIS_OUT_DIR="$projectHome/uc-non-persistent/test-results/analysis"
 
 infrastructureIds=(
-  "aws.devel1-standalone"
   "azure.devel1-standalone"
-  # "azure.devel2"
+  # "azure.devel2-standalone"
+  "aws.devel1-standalone"
+  # "aws.devel2-standalone"
 )
 
 export INFRASTRUCTURE_IDS="${infrastructureIds[*]}"