+++ uc-non-persistent/infrastructure/standalone/.devel/run.destroy.bg.sh
@@ -10,11 +10,13 @@ scriptName=$(basename $(test -L "$0" && readlink "$0" || echo "$0"));
 
 #  format: {cloud_provider}.{config}
 export infrastructureIds=(
-  # "aws.devel1"
+  "aws.devel1"
+  # "aws.devel2"
   "azure.devel1"
   # "azure.devel2"
 )
 
+export ANSIBLE_VERBOSITY=3
 
 export INFRASTRUCTURE_IDS="${infrastructureIds[*]}"