+++ uc-non-persistent/infrastructure/standalone/.test/run.destroy.fg.sh
@@ -11,9 +11,11 @@ scriptName=$(basename $(test -L "$0" && readlink "$0" || echo "$0"));
 #  format: {cloud_provider}.{config}
 export infrastructureIds=(
   "azure.test1"
-  # "azure.test2"
+  "azure.test2"
   "aws.test1"
 )
+
+export ANSIBLE_VERBOSITY=3
 
 export INFRASTRUCTURE_IDS="${infrastructureIds[*]}"