+++ uc-non-persistent/infrastructure/standalone/aws/aws-bootstrap.tf
@@ -30,12 +30,16 @@ resource "null_resource" "trigger_bootstrap" {
   }
   provisioner "local-exec" {
     # requires env var set: export ANSIBLE_PYTHON_INTERPRETER={path-to-python-3}
-    command = "../bootstrap/_run.bootstrap.sh aws.${var.tag_name_prefix}-standalone"
+    command = <<EOT
+      export APPLY_KERNEL_OPTIMIZATIONS=${var.apply_kernel_optimizations}
+      export APPLY_MELLANOX_VMA=${var.apply_mellanox_vma}
+      ../bootstrap/_run.bootstrap.sh aws.${var.tag_name_prefix}-standalone
+EOT
   }
   provisioner "local-exec" {
     when    = destroy
     # requires env var set: export ANSIBLE_PYTHON_INTERPRETER={path-to-python-3}
-    command = "../bootstrap/_run.bootstrap.destroy.sh azure.${self.triggers.tag_name_prefix}-standalone"
+    command = "../bootstrap/_run.bootstrap.destroy.sh aws.${self.triggers.tag_name_prefix}-standalone"
   }
   depends_on = [
       local_file.inventory_file