+++ uc-non-persistent/infrastructure/standalone/bootstrap/.devel/run.apply.kernel.optimizations.sh
@@ -7,23 +7,14 @@
 scriptDir=$(cd $(dirname "$0") && pwd);
 scriptName=$(basename $(test -L "$0" && readlink "$0" || echo "$0"));
 
+# export UC_NON_PERSISTENT_INFRASTRUCTURE="aws.devel1-standalone"
+export UC_NON_PERSISTENT_INFRASTRUCTURE="azure.devel1-standalone"
 
-#  format: {cloud_provider}.{config}
-export infrastructureIds=(
-  "azure.test1"
-  "azure.test2"
-  "azure.test3"
-  "azure.test4"
-)
-
-export INFRASTRUCTURE_IDS="${infrastructureIds[*]}"
-
+export ANSIBLE_VERBOSITY=3
 export LOG_DIR=$scriptDir/logs
 rm -f $LOG_DIR/*
 
-export TF_VARIABLES_DIR=$scriptDir
-
-nohup ../_run.apply-all.sh > $LOG_DIR/$scriptName.out 2>&1 &
+../_run.apply.kernel.optimizations.sh $UC_NON_PERSISTENT_INFRASTRUCTURE
 
 ###
 # The End.