+++ uc-non-persistent/infrastructure/standalone/bootstrap/.devel/run.bootstrap.sh
@@ -7,8 +7,11 @@
 scriptDir=$(cd $(dirname "$0") && pwd);
 scriptName=$(basename $(test -L "$0" && readlink "$0" || echo "$0"));
 
-export UC_NON_PERSISTENT_INFRASTRUCTURE="aws.devel1-standalone"
-# export UC_NON_PERSISTENT_INFRASTRUCTURE="azure.devel1-standalone"
+# export UC_NON_PERSISTENT_INFRASTRUCTURE="aws.devel1-standalone"
+export UC_NON_PERSISTENT_INFRASTRUCTURE="azure.devel1-standalone"
+
+export APPLY_KERNEL_OPTIMIZATIONS=false
+export APPLY_MELLANOX_VMA=true
 
 export ANSIBLE_VERBOSITY=3
 export LOG_DIR=$scriptDir/logs