+++ uc-non-persistent/infrastructure/standalone/bootstrap/.devel/run.remove.kernel.optimizations.sh
@@ -7,13 +7,14 @@
 scriptDir=$(cd $(dirname "$0") && pwd);
 scriptName=$(basename $(test -L "$0" && readlink "$0" || echo "$0"));
 
-export UC_NON_PERSISTENT_INFRASTRUCTURE="aws.devel1-standalone"
+# export UC_NON_PERSISTENT_INFRASTRUCTURE="aws.devel1-standalone"
+export UC_NON_PERSISTENT_INFRASTRUCTURE="azure.devel1-standalone"
 
 export ANSIBLE_VERBOSITY=3
 export LOG_DIR=$scriptDir/logs
 rm -f $LOG_DIR/*
 
-../_run.bootstrap.sh $UC_NON_PERSISTENT_INFRASTRUCTURE
+./_run.bootstrap.remove.kernel.optimizations.sh $UC_NON_PERSISTENT_INFRASTRUCTURE
 
 ###
 # The End.